chore(deps): Upgrade nextjs version.
As a part of the upgrade I finally managed to fix a bunch of
long-standing issues.

This PR also brings a huge rework to i18n thanks to some behind the
scene changes in nextjs 13.

Finally, currencies are finally fixed; visually. We still need to
properly integrate them with the Storefront API.