Meta: Improve `FIXME` and `TODO` comments.
In the future we should add a `FIXME` and `TODO` plugin+rules to
`eslint` to ensure they follow the following format:

```typescript
// FIXME: <description>.
// TODO: <description>.

/* FIXME: <description>. */
/* TODO: <description>. */

// FIXME: A `FIXME` description that is several lines long beacuse it
//        goes into detail about something that needs to be fixed.

// TODO: A `TODO` description that is several lines long beacuse it
//       goes into detail about something that needs to be fixed.
```

We should also add a linter to enforce ending comments with a period,
and start them with a capital letter. For example:

```typescript
// Hello world.
/* Hello world. */

// A comment that is several lines long, maybe it's describing a
// specification, a function, or a class, or maybe something completely
// different. Maybe it should be summarized into a single line. We don't
// know, but we do know that it should end with a period.

/***
 * Some description of a function.
 *
 * @param {string} arg1 - Some argument.
 * @returns {string} Some return value, reference to `arg1` argument.
 */
```

That is all for the future (FRO-53), for now we just improve the
existing instances of `FIXME` and `TODO` comments.