UI+Header: Improve responsiveness and fix FRO-50.
This is not the proper way to fix FRO-50, but we're going to most likely
redesign everything when we migrate to a zero-runtime style system, so
there's no point in doing it 100% correctly right now.

Hopefully https://github.com/NordcomInc/nordstar will be ready soon in
time for the huge style system switch, but that depends on how much time
I am able to allocate for it.

Additionally, this commit also improves the responsiveness and general
design of the `<SplitView/>` and the `<Gallery/>` components. They
should now both be ready for the `app router` release of the frontend
with the the knowledge that we're eventually going to redesign
everything as soon as realistically possible.