Locale: Add proper support for locales to `<Link/>`.
This commit adds proper support for locales to `<Link/>` component. It
also removes `generateStaticParams` from most pages and let's the layout
pre-generate only the default locale's pages. In the future we might
want to let some tenants pre-generate all locales' pages; but that is
really time consuming and costly when thinking about machine time on
both Vercel and GitHub Actions, so let's not right now :^)